working on card layout
