+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/devoncox/workspace/nss-day-cohort-43.github.io/src/App.js":"1","/Users/devoncox/workspace/nss-day-cohort-43.github.io/src/components/thanks/Thanks.js":"2"},{"size":8869,"mtime":1606255254784,"results":"3","hashOfConfig":"4"},{"size":2099,"mtime":1606255254781,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":1,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"1drxks5",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/devoncox/workspace/nss-day-cohort-43.github.io/src/App.js",["10"],"/Users/devoncox/workspace/nss-day-cohort-43.github.io/src/components/thanks/Thanks.js",[],{"ruleId":null,"fatal":true,"severity":2,"message":"11","line":4,"column":1},"Parsing error: Unexpected token\n\n  2 | import \"./scss/styles.scss\";\n  3 | import { NavBar } from './components/nav/NavBar'\n> 4 | <<<<<<< HEAD\n    | ^\n  5 | import { RSVPButton } from './components/nav/RSVP'\n  6 | \n  7 |"]
+[{"/Users/audrey/Workspace/cohort43/nss-day-cohort-43.github.io/src/components/cards/CardList.js":"1","/Users/audrey/Workspace/cohort43/nss-day-cohort-43.github.io/src/App.js":"2","/Users/audrey/Workspace/cohort43/nss-day-cohort-43.github.io/src/components/cards/Card.js":"3"},{"size":2832,"mtime":1606759877123,"results":"4","hashOfConfig":"5"},{"size":8543,"mtime":1606759152934,"results":"6","hashOfConfig":"5"},{"size":4118,"mtime":1606759814713,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":5,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"13s8l1x",{"filePath":"10","messages":"11","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/audrey/Workspace/cohort43/nss-day-cohort-43.github.io/src/components/cards/CardList.js",["14","15","16","17","18"],"/Users/audrey/Workspace/cohort43/nss-day-cohort-43.github.io/src/App.js",["19"],"/Users/audrey/Workspace/cohort43/nss-day-cohort-43.github.io/src/components/cards/Card.js",[],{"ruleId":"20","severity":1,"message":"21","line":4,"column":10,"nodeType":"22","messageId":"23","endLine":4,"endColumn":19},{"ruleId":"20","severity":1,"message":"24","line":4,"column":21,"nodeType":"22","messageId":"23","endLine":4,"endColumn":24},{"ruleId":"20","severity":1,"message":"25","line":4,"column":26,"nodeType":"22","messageId":"23","endLine":4,"endColumn":29},{"ruleId":"20","severity":1,"message":"26","line":12,"column":13,"nodeType":"22","messageId":"23","endLine":12,"endColumn":26},{"ruleId":"20","severity":1,"message":"27","line":12,"column":28,"nodeType":"22","messageId":"23","endLine":12,"endColumn":39},{"ruleId":"20","severity":1,"message":"28","line":5,"column":10,"nodeType":"22","messageId":"23","endLine":5,"endColumn":20},"no-unused-vars","'Container' is defined but never used.","Identifier","unusedVar","'Row' is defined but never used.","'Col' is defined but never used.","'filteredMates' is assigned a value but never used.","'setFiltered' is assigned a value but never used.","'RSVPButton' is defined but never used."]
