+++ src/components/cards/CardList.js
@@ -1,52 +1,93 @@
+// By Audrey Thomasson
 
-
-/**
- * Temporary code to show both card types.
- * Replace with actual CardList Code
- */
-
-import React from 'react';
-import { CardHTML } from "./Card.js";
+import React, { useEffect, useState } from "react"
+import { Container, Row, Col } from 'react-bootstrap';
+import { CardHTML } from "./Card"
 import './customCard.scss';
 
+export const CardList = () => {
+    const [allMates, setAllMates] = useState([])
 
-export const TempCardList = () => {
-    const exampleDatabase = [
-        {
-            firstName: "Rick",
-            lastName: "Blake",
-            photo: "images/placeholder-images/BlakeRick.jpeg",
-            focus: "Full-Stack Developer",
-            linkedIn: "https://www.linkedin.com/in/blaker814/",
-            github: "https://github.com/blaker814",
-            personalSite: "",
-            dribbble: "",
-            id: 2
-        },
-        {
-            firstName: "Devon",
-            lastName: "Cox",
-            photo: "images/placeholder-images/CoxDevon.jpeg",
-            focus: "UI/UX Front-End Developer",
-            linkedIn: "https://www.linkedin.com/in/devonhcox/",
-            github: "https://github.com/Devco2011",
-            personalSite: "eeeeeee",
-            dribbble: "https://dribbble.com/DevCo",
-            id: 3
-        }
-    ]
+    // allMates sorted by name
+    const [ filteredMates, setFiltered ] = useState([])
 
-    const tempCardListHTML = exampleDatabase.map(student => {
-        return (
-            <>
-                <div className="col-md-6 col-lg-4 mb-5">
-                    {
-                        CardHTML(student)
-                    }
-                </div>
-            </>   
-        )
-    })
+    // function to make call to database to get all the classmates
+    const getAllMates = () => {
+        return fetch ("database.json")
+        	.then(response => response.json())
+    }
+    
+    // setting all the classmates
+	useEffect(() => {
+		getAllMates()
+			.then(data => {
+                setAllMates(filterNames(data.cohortMates))
+				
+			})
+	}, [])
 
-    return tempCardListHTML
+    // sort by name to create filtered list of all classmates
+
+    const filterNames = (allMates) => {
+        const alphaMates = allMates.sort(function(a, b) {
+        const nameA = a.firstName.toUpperCase(); // ignore upper and lowercase
+        const nameB = b.firstName.toUpperCase(); // ignore upper and lowercase
+            if (nameA < nameB) {
+            return -1;
+            }
+            if (nameA > nameB) {
+            return 1;
+            }
+            // names must be equal
+            return 0;
+        })
+        return alphaMates
+    }
+   
+
+    // // Filters for searchTerms and finds the classmate by first OR last name
+    // // searchTerms will cause a change
+    // useEffect(() => {
+    //     if (searchTerms !== "") {
+    //         // If the search field is not blank, display matching shots
+    //         const subset = mates.filter(mate => mate.firstName.toLowerCase().includes(searchTerms.toLowerCase()) || mate.lastName.toLowerCase().includes(searchTerms.toLowerCase()))
+    //         setFiltered(subset)
+    //     } else {
+    //         // If the search field is blank, display all shots matching the activeUser
+    //         setFiltered(mates)
+    //     }
+    // }, [searchTerms, mates])
+
+ 
+
+
+    return (
+        <>  
+            <div className="col-md-6 col-lg-4 mb-5">                    
+                {
+                    allMates.map(mate => {
+                        return <CardHTML key={mate.id} mate={mate} />
+                    })
+                }
+            </div>
+            
+            {/* <Container fluid>           
+                <Row>
+                    <Col xs={12} s={12} md={6} l={4} xl={4}>
+                        <>
+                            {
+                                allMates.map(mate => {
+                                    console.log(mate)
+                                    return <CardHTML key={mate.id} mate={mate} />
+                                })
+                            }
+                        </>
+                    </Col>
+                </Row>
+
+            </Container>  */}
+
+
+        </>
+    )
 }
